+++ src/laser_generic/models/shared.py
@@ -9,7 +9,7 @@ __all__ = ["State", "sample_dobs", "sample_dods"]
 
 def sample_dobs(dobs: np.ndarray, pyramid: AliasedDistribution, tick: int) -> None:
     # Get years of age sampled from the population pyramid
-    dobs[:] = pyramid.sample(len(dobs)).astype(np.int16)  # Fit in np.int16
+    dobs[:] = pyramid.sample(len(dobs)).astype(np.int32)  # Fit in np.int16
     dobs *= 365  # Convert years to days
     dobs += np.random.randint(0, 365, size=len(dobs))  # add some noise within the year
     # pyramid.sample actually returned ages. Turn them into dobs by treating them
@@ -21,10 +21,10 @@ def sample_dobs(dobs: np.ndarray, pyramid: AliasedDistribution, tick: int) -> No
 
 def sample_dods(dobs: np.ndarray, dods: np.ndarray, survival: KaplanMeierEstimator, tick: int) -> None:
     # An agent's age is (tick - dob).
-    dods[:] = survival.predict_age_at_death(ages := tick - dobs).astype(np.int16)  # Fit in np.int16
+    dods[:] = survival.predict_age_at_death(ages := tick - dobs).astype(np.int32)  # Fit in np.int16
     dods -= ages  # How many more days will each person live?
     dods += tick  # Convert to date of death
-    
+
     return
